requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
async-timeout==4.0.3
attrs==23.1.0
blinker==1.7.0
cattrs==22.2.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
exceptiongroup==1.1.2
Flask==3.0.2
frozenlist==1.4.0
highrise-bot-sdk==23.3.4
idna==3.7
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.5
multidict==6.0.4
pendulum==2.1.2
python-dateutil==2.8.2
python-dotenv==1.0.1
pytzdata==2020.1
quattro==22.2.0
requests==2.31.0
six==1.16.0
typing-extensions==3.10.0.2
urllib3==2.2.1
Werkzeug==3.0.1
yarl==1.9.2
